Make README include work for both md and rst and refactor
When using MyST to include Markdown-READMEs a different include command
is needed in case of a README in rst format.  Therefore move the code
for searching and copying the README into a function and construct the
include command there, depending on the format of the file.
This is then added to the index.rst using a template variable.

Move the equivalent code for the license file to its own function in the
same way, to be consistent.