Allow plain text READMEs
Also check for "readme" and "readme.txt" (case insensitive).  If one of
these is found, include it with `:literal:`.

If multiple README files are found, simply use the first one found and
print a warning.  Normally a package should not have multiple READMEs so
if this is the case it should better be fixed by the user.