Use variconf to find and load config files
Using variconf makes loading the config file much easier, as we don't
have to reinvent the wheel for handling default values.

Unfortunately, the current syntax of using `${PACKAGE_DIR}` in Doxygen
exclude patterns collides with OmegaConf's variable interpolation
feature.  Therefore, change the syntax in the exclude patterns to
`{{PACKAGE_DIR}}`.

Using `${PACKAGE_DIR}` is now deprecated but currently still supported
(so this is not a breaking change) via a little hack: Adding
`"PACKAGE_DIR": "{{PACKAGE_DIR}}"` to the default configuration let's
OmegaConf replace `${PACKAGE_DIR}` with `{{PACKAGE_DIR}}` when loading
the file.

Moved the code for constructing the EXCLUDE_PATTERNS string into its own
function, so it is easier to test.