fix: Make config non-strict to allow intersphinx mappings
The strict config was not allowing the flexible use of either a string
or a dictionary in the intersphinx mapping settings (unfortunately
unions are not fully supported by OmegaConf).
To fix this for now, make the config non-strict (this means no error
will be triggered anymore if the user has a typo in one of the config
parameter names).