Use index.rst in package root as alternative to README
Add "index.rst" as a candidate name for the README file (rename the
corresponding functions to "..._index_or_readme" to reflect this).

Rewrite the code search for the file such that in case of multiple
matches, the order in which they are specified in `options` is adhered
to (previously it was just randomly picking whichever was first returned
by `glob("*")`).

It resolves the issue that the content in the README does not always
make sense to be used as documentation main page by adding the option to
provide an alternative index.rst next to the README (GitHub will still
show the README but the documentation will use the index.rst).